a very big commit to rework all directories and allow for more trip pages
