+++ app/routes/__init__.py
@@ -1,0 +1,1 @@
+# This file can be empty - it just marks the directory as a Python package
