+++ app/static/css/styles/animations/_animations.css
