+++ app/static/css/styles/components/_trips.css
