+++ app/static/script.js
